refactor: finish extracting ListTablePage component
